requirements-dev.txt
black
django-debug-toolbar>=6.0.0
django-stubs>=5.2.9,<5.3
flake8==7.3.0
flake8-pyproject==1.2.4
isort
mock==5.2.0
mypy>=1.19.1
mypy_extensions
pre-commit>=4.5.1
setuptools>=80.10.2
